Check if no swap

+++ packer/centos6/scripts/zerodisk.sh
@@ -4,11 +4,17 @@
 echo "[*] Zeroing free space"
 
 swapuuid=`blkid -o value -l -s UUID -t TYPE=swap`
-swappart=`readlink -f /dev/disk/by-uuid/$swapuuid`
 
-swapoff $swappart
-dd if=/dev/zero of=$swappart bs=1M
-mkswap -U $swapuuid $swappart
+if [ -n "$swapuuid" ]
+then
+  swappart=`readlink -f /dev/disk/by-uuid/$swapuuid`
+
+  swapoff $swappart
+  dd if=/dev/zero of=$swappart bs=1M
+  mkswap -U $swapuuid $swappart
+else
+  echo "[*] No swap detected"
+fi
 
 dd if=/dev/zero of=/EMPTY bs=1M
 rm -f /EMPTY
